+++ sdk/ai/azure-ai-projects/samples/evaluation/sample_evaluations_async.py
@@ -98,7 +98,7 @@ async def main() -> None:
                 evaluation,
                 headers={
                     "model-endpoint": model_endpoint,
-                    "api-key": model_api_key,
+                    "model-api-key": model_api_key,
                 },
             )
             print(evaluation_response)