+++ sdk/computeschedule/azure-mgmt-computeschedule/generated_samples/scheduled_actions_virtual_machines_submit_deallocate_maximum_set_gen.py
@@ -32,30 +32,28 @@ def main():
     )
 
     response = client.scheduled_actions.virtual_machines_submit_deallocate(
-        locationparameter="smdt",
+        locationparameter="ycipx",
         request_body={
-            "correlationid": "evmwonebfzxenjdpucgcwdjdya",
+            "correlationid": "b211f086-4b91-4686-a453-2f5c012e4d80",
             "executionParameters": {
                 "optimizationPreference": "Cost",
-                "retryPolicy": {"retryCount": 25, "retryWindowInMinutes": 4},
+                "retryPolicy": {"retryCount": 17, "retryWindowInMinutes": 29},
             },
             "resources": {
                 "ids": [
-                    "/subscriptions/YourSubscriptionId/resourceGroups/YourResourceGroupName/providers/Microsoft.Compute/virtualMachines/testResource3"
+                    "/subscriptions/YourSubscriptionId/resourceGroups/YourResourceGroupName/providers/Microsoft.Compute/virtualMachines/testResource4"
                 ]
             },
             "schedule": {
-                "deadLine": "2025-04-15T19:47:04.403Z",
-                "deadline": "2025-04-15T19:47:04.403Z",
+                "deadLine": "2025-04-17T00:23:56.803Z",
                 "deadlineType": "Unknown",
-                "timeZone": "upnmayfebiadztdktxzq",
-                "timezone": "qacufsmctpgjozovlsihrzoctatcsj",
+                "timeZone": "aigbjdnldtzkteqi",
             },
         },
     )
     print(response)
 
 
-# x-ms-original-file: 2025-05-01/ScheduledActions_VirtualMachinesSubmitDeallocate_MaximumSet_Gen.json
+# x-ms-original-file: 2025-04-15-preview/ScheduledActions_VirtualMachinesSubmitDeallocate_MaximumSet_Gen.json
 if __name__ == "__main__":
     main()