+++ sdk/containerservice/azure-mgmt-containerservice/generated_samples/managed_clusters_create_gpumig.py
@@ -77,6 +77,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/containerservice/resource-manager/Microsoft.ContainerService/aks/stable/2025-04-01/examples/ManagedClustersCreate_GPUMIG.json
+# x-ms-original-file: specification/containerservice/resource-manager/Microsoft.ContainerService/aks/stable/2025-05-01/examples/ManagedClustersCreate_GPUMIG.json
 if __name__ == "__main__":
     main()