+++ sdk/eventgrid/azure-mgmt-eventgrid/generated_tests/test_event_grid_management_topic_spaces_operations.py
@@ -25,7 +25,7 @@ class TestEventGridManagementTopicSpacesOperations(AzureMgmtRecordedTestCase):
             resource_group_name=resource_group.name,
             namespace_name="str",
             topic_space_name="str",
-            api_version="2025-02-15",
+            api_version="2025-04-01-preview",
         )
 
         # please add some check logic here by yourself
@@ -54,7 +54,7 @@ class TestEventGridManagementTopicSpacesOperations(AzureMgmtRecordedTestCase):
                 "topicTemplates": ["str"],
                 "type": "str",
             },
-            api_version="2025-02-15",
+            api_version="2025-04-01-preview",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -67,7 +67,7 @@ class TestEventGridManagementTopicSpacesOperations(AzureMgmtRecordedTestCase):
             resource_group_name=resource_group.name,
             namespace_name="str",
             topic_space_name="str",
-            api_version="2025-02-15",
+            api_version="2025-04-01-preview",
         ).result()  # call '.result()' to poll until service return final result
 
         # please add some check logic here by yourself
@@ -79,7 +79,7 @@ class TestEventGridManagementTopicSpacesOperations(AzureMgmtRecordedTestCase):
         response = self.client.topic_spaces.list_by_namespace(
             resource_group_name=resource_group.name,
             namespace_name="str",
-            api_version="2025-02-15",
+            api_version="2025-04-01-preview",
         )
         result = [r for r in response]
         # please add some check logic here by yourself