+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 1.0.0b40 (Unreleased)
+## 1.0.0b41 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0b40 (2025-07-17)
 
 ### Features Added
 
@@ -10,12 +20,14 @@
   ([#41669](https://github.com/Azure/azure-sdk-for-python/pull/41669))
 - Customer Facing Statsbeat: Added logic for dropped item count
   ([#41950](https://github.com/Azure/azure-sdk-for-python/pull/41950))
+- Customer Facing Statsbeat: Added logic for retry item count
+  ([#41971](https://github.com/Azure/azure-sdk-for-python/pull/41971))
+- Added RateLimited Sampler 
+  ([#41954](https://github.com/Azure/azure-sdk-for-python/pull/41954))
+- Refactored Application Insights Sampler Code
+  ([#42210](https://github.com/Azure/azure-sdk-for-python/pull/42210))
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
+- Support AI Foundry by Handling GEN_AI_SYSTEM Attributes with [Spec](https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/blob/main/ApplicationInsights/genai_semconv_mapping.md) ([#41705](https://github.com/Azure/azure-sdk-for-python/pull/41705))
 
 ## 1.0.0b39 (2025-06-25)